完成了除链表外的其他题目
